Move Page definition objects into separate trait
This makes it easier to keep track of pages and makes the testing logic
easier to read.
